requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
langchain-openai
sentence-transformers
lightrag-hku
mineru[core]
Pillow
pymupdf
